Refactor parseUri method into diag-data module
