+++ AngularApp/projects/app-service-diagnostics/src/app/home/components/home/home.component.ts
@@ -1,4 +1,4 @@
-import { DetectorControlService, FeatureNavigationService, DetectorResponse, TelemetryEventNames } from 'diagnostic-data';
+import { DetectorControlService, FeatureNavigationService, DetectorResponse, TelemetryEventNames, ResourceDescriptor } from 'diagnostic-data';
 import { Component, OnInit, AfterViewInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Category } from '../../../shared-v2/models/category';
@@ -200,14 +200,14 @@ export class HomeComponent implements OnInit, AfterViewInit {
     }
 
    
-  public get useStaticAksText() : boolean {
-    return this.armService.isMooncake 
-    && this._resourceService.parseResourceUri(this._resourceService.resourceIdForRouting).provider.toLowerCase() == 'microsoft.containerservice';
+    public get useStaticAksText() : boolean {
+        return this.armService.isMooncake 
+            && ResourceDescriptor.parseResourceUri(this._resourceService.resourceIdForRouting).provider.toLowerCase() == 'microsoft.containerservice';
   }
 
     public get isAKSOnNationalCloud(): boolean {
         return this.armService.isNationalCloud
-            && this._resourceService.parseResourceUri(this._resourceService.resourceIdForRouting).provider.toLowerCase() == 'microsoft.containerservice';
+            && ResourceDescriptor.parseResourceUri(this._resourceService.resourceIdForRouting).provider.toLowerCase() == 'microsoft.containerservice';
     }
 
     onSearchBoxFocus(event: any): void {
