+++ AngularApp/projects/app-service-diagnostics/src/app/shared/models/arm/armResourceConfig.ts
@@ -28,30 +28,3 @@ export class HomePageText {
 	description: string;
 	searchBarPlaceHolder: string;
 }
-
-export enum ResourceDescriptorGroups {
-	subscription = 1,
-	resourceGroup = 3,
-	provider = 4,
-	resource = 5
-}
-export class ResourceDescriptor {
-	constructor() {
-		this.subscription = '';
-		this.resourceGroup = '';
-		this.provider = '';
-		this.type = '';
-		this.resource = '';
-		this.types = [];
-		this.resources = [];
-		this.resourceUriRegExp = new RegExp('/subscriptions/([^/]+)/(resourceGroups/([^/]+)/)?providers/([^/]+)/(.+)', "i");
-	}
-	subscription: string;
-	resourceGroup: string;
-	provider: string;
-	type: string;
-	resource: string;
-	types: Array<string>;
-	resources: Array<string>;
-	resourceUriRegExp:RegExp;
-}
